Support specific glibc versions
It is now possible to choose the glibc version to link against.
Note that for now, we use the same set of headers regarless of the
glibc version.

It was also decided to generate all libs of the glibc even if
they are integrated inside libc in newer versions of glibc (like
-lm -rt etc.). This simplifies conditional linking a lot.