Support stack smashing protector symbols from the dynamic linker
__stack_chk_guard is provided by the dynamic linker but we can't
link against it for the moment. We can only set it's correct path
depending on the target architecture thanks to the clang frontend.

But because we don't link against it, code using ssp symbols will
fail because __stack_chk_guard remains undefined.

Here, we add it as a stub in the libc knowing that it will be
provided by the dynamic linker.