Allow disable empty sysroot
This is needed when linking rust on linux.
This is because the rust toolchain uses prebuilt rlibs that embed
link flag propagation information, and some link flags are expected
to exist.
This is unfortunate because it means that the decision of libraries
to link against has been taken ahead of time...

See https://github.com/rust-lang/rust/blob/ec28ae9454139023117270985f114823d6570657/library/unwind/src/lib.rs#L107-L123
which makes prebuilt libunwind in the rust toolchain propagate
-lgcc_s in all situations...