Add a build_setting to allow stub linking against libgcc_s
This flag makes a dummy gcc_s library to link against.

libgcc_s is a shared library (only libgcc_s.so exists) that is required
when creating or linking against a shared library that uses c++ exceptions
that may cross the library boundary.

This toolchain currently doesn't support linking dynamically against an
unwinder, which means that this toolchain doesn't support cross boundary
c++ exceptions for the moment (and the only unwinder supported is libunwind).
Yet, it is possible for dependencies that you do not control to pass -lgcc_s
linker flags.

If you know what you are doing and want to no-op these flags, this option
will make a stub libgcc_s to satistfy the linker, but will still link libunwind
statically.

In theory, such option should only be available if this toolchain had support
for a dynamically linked unwinder.