Use a link only resource-dir (#125)
This is primarily so that go tool link linkerFlagSupported checks, which
trims a lot of flags, can pass.
But actually this is also the best strategy for our toolchain for now,
because we should not mix half driver, half not driver.
Additionally, this will also make it easier to generate cross toolchains
for usage outside of bazel.