Stop using a dedicated cc_toolchain for stage2 (#162)
We now use the transition to choose the correct toolchain args within a
single toolchain.
I also took the opportunity to clean a bit the args names and usages.

I want to do a bigger revamp of whole naming and fix actualy default
args but thought I'd do that in multiple passes !