Restructure toolchain target settings (#171)
The idea is to simplify the mental model of toolchain resolution,
especially as we bring "stage1" (to be renamed).

In this PR:
1. We change config settings names to better match the intents
2. We change the matching strategy to be explicit vs fallback

For 1:

Which runtimes do we need? (no runtimes? + CRTs? +libc? +libc++?)
Which compiler should be used? (prebuilt? from source aka bootstrapped?)

For 2:

We start with default build settings value for both:
- all runtimes
- prebuilt

And we constraint toolchain declarations to either prebuilt or
bootstrapped.

This allows regstering all toolchains, and using stage1 by using a build
flag.
```
common:stage1   --@toolchains_llvm_bootstrapped//toolchain:compiler_flavor=bootstrapped
```